requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cssselect==1.3.0
dateparser==1.2.1
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
executing==2.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba3k==0.35.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.1
lxml_html_clean==0.4.1
matplotlib-inline==0.1.7
multidict==6.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pygooglenews==0.1.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
selenium==4.29.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tinysegmenter==0.3
tldextract==5.1.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.54
